LayoutTests/imported/w3c/web-platform-tests/streams/writable-streams/start.any-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: undefined
CONSOLE MESSAGE: Unhandled Promise Rejection: [object Object]
CONSOLE MESSAGE: Unhandled Promise Rejection: undefined

Harness Error (FAIL), message = Unhandled rejection

FAIL underlying sink's write should not be called until start finishes ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)
FAIL underlying sink's close should not be called until start finishes ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)
PASS underlying sink's write or close should not be called if start throws 
PASS underlying sink's write or close should not be invoked if the promise returned by start is rejected 
FAIL returning a thenable from start() should work ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)
FAIL controller.error() during start should cause writes to fail controller.error is not a function. (In 'controller.error(error1)', 'controller.error' is undefined)
FAIL controller.error() during async start should cause existing writes to fail ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)
FAIL when start() rejects, writer promises should reject in standard order ws.getWriter is not a function. (In 'ws.getWriter()', 'ws.getWriter' is undefined)
